refactor: move json feed rendering to separated module (#11938)
